--- lib/external-types.d.ts
@@ -1,3 +1,0 @@
-declare module 'dashify' {
-  export default function(input: string): string;
-}
