--- lib/index.ts
@@ -1,9 +1,0 @@
-'use strict';
-
-export { default as buildAttr } from './-private/build-attr';
-export { default as appendToContent } from './-private/append-to-content';
-export {
-  default as BuildTimeComponent,
-  BuildTimeComponentNode,
-  BuildTimeComponentOptions
-} from './-private/build-time-component';
