+++ test/unit/append-to-content-test.ts
@@ -2,14 +2,14 @@
 
 import { builders as b } from '@glimmer/syntax';
 import processTemplate from '../helpers/process-template';
-import { appendToContent } from '../../lib';
+import { appendToAttrContent } from '../../lib/html';
 
-describe('Helper #appendToContent', function() {
+describe('Helper #appendToAttrContent', function() {
   it('it can append undefined (which does nothing)', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(undefined, attr.value);
-        attr.value = appendToContent(undefined, attr.value);
+        attr.value = appendToAttrContent(undefined, attr.value);
+        attr.value = appendToAttrContent(undefined, attr.value);
         return attr;
       }
     });
@@ -20,8 +20,8 @@ describe('Helper #appendToContent', function() {
   it('it can append null (which does nothing)', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(null, attr.value);
-        attr.value = appendToContent(null, attr.value);
+        attr.value = appendToAttrContent(null, attr.value);
+        attr.value = appendToAttrContent(null, attr.value);
         return attr;
       }
     });
@@ -32,8 +32,8 @@ describe('Helper #appendToContent', function() {
   it('it can append regular strings', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent('bar', attr.value);
-        attr.value = appendToContent('baz', attr.value);
+        attr.value = appendToAttrContent('bar', attr.value);
+        attr.value = appendToAttrContent('baz', attr.value);
         return attr;
       }
     });
@@ -44,8 +44,8 @@ describe('Helper #appendToContent', function() {
   it('it can append StringLiterals', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.string('bar'), attr.value);
-        attr.value = appendToContent(b.string('baz'), attr.value);
+        attr.value = appendToAttrContent(b.string('bar'), attr.value);
+        attr.value = appendToAttrContent(b.string('baz'), attr.value);
         return attr;
       }
     });
@@ -56,8 +56,8 @@ describe('Helper #appendToContent', function() {
   it('it can append NumberLiterals', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.number(1), attr.value);
-        attr.value = appendToContent(b.number(2), attr.value);
+        attr.value = appendToAttrContent(b.number(1), attr.value);
+        attr.value = appendToAttrContent(b.number(2), attr.value);
         return attr;
       }
     });
@@ -68,8 +68,8 @@ describe('Helper #appendToContent', function() {
   it('it can append TextNodes to another TextNode', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.text('bar'), attr.value);
-        attr.value = appendToContent(b.text('baz'), attr.value);
+        attr.value = appendToAttrContent(b.text('bar'), attr.value);
+        attr.value = appendToAttrContent(b.text('baz'), attr.value);
         return attr;
       }
     });
@@ -80,8 +80,8 @@ describe('Helper #appendToContent', function() {
   it('it can append PathExpression', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.path('bar'), attr.value);
-        attr.value = appendToContent(b.path('baz'), attr.value);
+        attr.value = appendToAttrContent(b.path('bar'), attr.value);
+        attr.value = appendToAttrContent(b.path('baz'), attr.value);
         return attr;
       }
     });
@@ -92,9 +92,9 @@ describe('Helper #appendToContent', function() {
   it('it can append MustacheStatement', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.mustache(b.path('bar')), attr.value);
-        attr.value = appendToContent(b.mustache(b.path('baz')), attr.value);
-        attr.value = appendToContent(b.mustache(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value);
+        attr.value = appendToAttrContent(b.mustache(b.path('bar')), attr.value);
+        attr.value = appendToAttrContent(b.mustache(b.path('baz')), attr.value);
+        attr.value = appendToAttrContent(b.mustache(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value);
         return attr;
       }
     });
@@ -105,8 +105,8 @@ describe('Helper #appendToContent', function() {
   it('it can append SubExpressions', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.sexpr(b.path('some-helper'), [b.string('someArg')]), attr.value);
-        attr.value = appendToContent(b.sexpr(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value);
+        attr.value = appendToAttrContent(b.sexpr(b.path('some-helper'), [b.string('someArg')]), attr.value);
+        attr.value = appendToAttrContent(b.sexpr(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value);
         return attr;
       }
     });
@@ -118,9 +118,9 @@ describe('Helper #appendToContent', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
         let val = b.concat([b.text('prefix'), b.mustache(b.path('boundVal')), b.text('suffix')]);
-        attr.value = appendToContent(val, attr.value);
+        attr.value = appendToAttrContent(val, attr.value);
         let val2 = b.concat([b.text('prefix2'), b.mustache(b.path('boundVal2'))]);
-        attr.value = appendToContent(val2, attr.value);
+        attr.value = appendToAttrContent(val2, attr.value);
         return attr;
       }
     });
@@ -131,14 +131,14 @@ describe('Helper #appendToContent', function() {
   it('it can mix and append a mix of elements', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.mustache(b.path('one')), b.text(''));
-        attr.value = appendToContent(b.mustache(b.path('two')), attr.value);
-        attr.value = appendToContent(b.text('three'), attr.value);
-        attr.value = appendToContent('four', attr.value);
-        attr.value = appendToContent(b.path('five'), attr.value);
-        attr.value = appendToContent(b.mustache(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value);
-        attr.value = appendToContent(b.path('six'), attr.value);
-        attr.value = appendToContent(b.concat([b.mustache(b.path('sev')), b.text('en')]), attr.value);
+        attr.value = appendToAttrContent(b.mustache(b.path('one')), b.text(''));
+        attr.value = appendToAttrContent(b.mustache(b.path('two')), attr.value);
+        attr.value = appendToAttrContent(b.text('three'), attr.value);
+        attr.value = appendToAttrContent('four', attr.value);
+        attr.value = appendToAttrContent(b.path('five'), attr.value);
+        attr.value = appendToAttrContent(b.mustache(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value);
+        attr.value = appendToAttrContent(b.path('six'), attr.value);
+        attr.value = appendToAttrContent(b.concat([b.mustache(b.path('sev')), b.text('en')]), attr.value);
         return attr;
       }
     });
@@ -149,14 +149,14 @@ describe('Helper #appendToContent', function() {
   it('it can mix and append a mix of elements without prepending spaces', function() {
     let modifiedTemplate = processTemplate(`<div class="foo"></div>`, {
       AttrNode(attr) {
-        attr.value = appendToContent(b.mustache(b.path('one')), b.text(''), { prependSpace: false });
-        attr.value = appendToContent(b.mustache(b.path('two')), attr.value, { prependSpace: false });
-        attr.value = appendToContent(b.text('three'), attr.value, { prependSpace: false });
-        attr.value = appendToContent('four', attr.value, { prependSpace: false });
-        attr.value = appendToContent(b.path('five'), attr.value, { prependSpace: false });
-        attr.value = appendToContent(b.mustache(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value, { prependSpace: false });
-        attr.value = appendToContent(b.path('six'), attr.value, { prependSpace: false });
-        attr.value = appendToContent(b.concat([b.mustache(b.path('sev')), b.text('en')]), attr.value, { prependSpace: false });
+        attr.value = appendToAttrContent(b.mustache(b.path('one')), b.text(''), { prependSpace: false });
+        attr.value = appendToAttrContent(b.mustache(b.path('two')), attr.value, { prependSpace: false });
+        attr.value = appendToAttrContent(b.text('three'), attr.value, { prependSpace: false });
+        attr.value = appendToAttrContent('four', attr.value, { prependSpace: false });
+        attr.value = appendToAttrContent(b.path('five'), attr.value, { prependSpace: false });
+        attr.value = appendToAttrContent(b.mustache(b.path('if'), [b.path('condition'), b.string('yes'), b.string('no')]), attr.value, { prependSpace: false });
+        attr.value = appendToAttrContent(b.path('six'), attr.value, { prependSpace: false });
+        attr.value = appendToAttrContent(b.concat([b.mustache(b.path('sev')), b.text('en')]), attr.value, { prependSpace: false });
         return attr;
       }
     });
