+++ test/unit/build-attr-test.ts
@@ -2,7 +2,7 @@
 
 import { builders as b, AST } from '@glimmer/syntax';
 import processTemplate from '../helpers/process-template';
-import { buildAttr } from '../../lib';
+import { buildAttr } from '../../lib/html';
 
 describe('Helper #appendToContent', function() {
   it('it builds attrs given a string', function() {
